% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_depth_to_bytes.R
\name{batch_depth_to_bytes}
\alias{batch_depth_to_bytes}
\title{Convert postage batch depth to bytes}
\usage{
batch_depth_to_bytes(depth, scale_in = "GB")
}
\arguments{
\item{depth}{Depth of the postage batch.}

\item{scale_in}{Scale to convert to - "GB" or "TB" for Gigabytes and Terabytes.}
}
\value{
Returns value in specified multiples of bytes.
}
\description{
Function converts postage batch depth into more human readable units of byte multiples.
}
