% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_price_of_storage.R
\name{batch_price_of_storage}
\alias{batch_price_of_storage}
\title{Calculate postage batch price for depth and duration}
\usage{
batch_price_of_storage(depth, duration_in_blocks = 17280, price_per_chunk)
}
\arguments{
\item{depth}{Depth of the postage batch.}

\item{duration_in_blocks}{Duration of the postage stamp in blocks on Gnosis chain.
Assuming each block happens on average in 5 sec on xDAI, there are approx. 60 * 60 * 24 / 5 = 17280 blocks per day.}

\item{price_per_chunk}{The price of storing one chunk for one block, in BZZ (PLUR).}
}
\value{
Returns value in xBZZ.
}
\description{
Function postage batch price in xBZZ, given the depth and duration.
}
