% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dec_to_bin.R
\name{dec2bin}
\alias{dec2bin}
\title{Convert decimal to binary}
\usage{
dec2bin(x)
}
\description{
Helper function, as per https://stackoverflow.com/questions/6614283/converting-decimal-to-binary-in-r
without rev()
}
