% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/first_n_places.r
\name{first_n_places}
\alias{first_n_places}
\title{Return first n characters of a string as string}
\usage{
first_n_places(string, n)
}
\arguments{
\item{string}{The string to process.}

\item{n}{The number of characters to return.}
}
\value{
Returns first n characters as a string.
}
\description{
Return first n characters of a string as string
}
