% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_short_overlays.R
\name{generate_short_overlay}
\alias{generate_short_overlay}
\title{Generate neighbourhood overlay bits for some radius}
\usage{
generate_short_overlay(radius)
}
\arguments{
\item{radius}{Storage radius.}
}
\value{
Returns overlays.
}
\description{
Generates all possible neighbourhod binary addresses for some storage radius, to use as factor variable.
}
