% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_last_round.R
\name{get_last_round}
\alias{get_last_round}
\title{Get last played redistribution round from Swarmscan}
\usage{
get_last_round()
}
\value{
Returns number of last (latest) played round
}
\description{
Function that call the Swarmscan.io API to get a list of played rounds and
extracts the latest (last) played round.
}
