% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hexadecimal_to_binary.R
\name{hexadecimal2binary}
\alias{hexadecimal2binary}
\title{Convert any length hex number to binary, in string representation}
\usage{
hexadecimal2binary(hexNumber)
}
\arguments{
\item{hexNumber}{The hex address of a Bee node.}
}
\value{
Returns a binary representation, as a string.
}
\description{
Takes a hex address of a Bee node and converts it to binary representation.
}
