% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_global_stamp_data.R
\name{load_global_stamp_data}
\alias{load_global_stamp_data}
\title{Load global stamp data from a Bee endpoint}
\usage{
load_global_stamp_data()
}
\value{
Returns a data frame with data about stamps.
}
\description{
Function to load global stamp data from local Bee, debug API; prerequisite: a running local Bee.
}
