% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_swarmscan_countries.R
\name{load_swarmscan_countries}
\alias{load_swarmscan_countries}
\title{Load swarmscan countries data}
\usage{
load_swarmscan_countries()
}
\value{
Returns a data frame parsed from JSON response.
}
\description{
Loads data from swarmscan.io about countries from endpoint "https://api.swarmscan.io/v1/network/stats".
}
