% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_swarmscan_data.R
\name{load_swarmscan_data}
\alias{load_swarmscan_data}
\title{Load swarmscan nodes data}
\usage{
load_swarmscan_data()
}
\value{
Returns a data frame parsed from the JSON response.
}
\description{
Loads swarmscan.io data about nodes on the network, from the "https://api.swarmscan.io/v1/network/dump" endpoint.
}
