% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_swarmscan_staking.R
\name{load_swarmscan_staking}
\alias{load_swarmscan_staking}
\title{Load staking swarmscan data}
\usage{
load_swarmscan_staking()
}
\value{
Returns a dataframe.
}
\description{
Loads data about staking from swarmscan.io, using endpoint "https://api.swarmscan.io/v1/events/staking/stake-updated/dump".
}
