% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nodes_for_storage.R
\name{nodes_needed_for_storage}
\alias{nodes_needed_for_storage}
\title{Plot storage bought per day}
\usage{
nodes_needed_for_storage(
  storageAmount,
  unit = "GB",
  reserveSizeChunks = (2^22),
  minNodesInNbhood = 4
)
}
\arguments{
\item{storageAmount}{Amount of storage required, numeric.}

\item{unit}{The unit for the amount, "GB" or "TB"; defaults to "GB".}

\item{reserveSizeChunks}{The amount of chunks held in the reserve of a Bee node, defaults to cca. 4 Mio (2 ^ 22); leave it as it is unless you know what you are doing.}

\item{minNodesInNbhood}{How many nodes should be in each neighbourhood. Defaults to 4 (appropriate minimal redundancy, but assuming equal distribution)}
}
\value{
Returns the number of neighbourhoods and nodes needed to ensure the amount of storage, based on parameters.
}
\description{
Function to plot a graph of storage paid for per day.
}
