% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/number_of_neighbourhoods.R
\name{numNbhoodsHigherThan}
\alias{numNbhoodsHigherThan}
\title{Round up number of neighbourhoods needed to a nearest power of 2}
\usage{
numNbhoodsHigherThan(minNhoodNumber)
}
\arguments{
\item{minNhoodNumber}{Minimum number of neighbourhoods needed.}
}
\value{
Returns a rounded up number of neighbourhoods.
}
\description{
Function round up number of neighbourhoods needed, based on power of 2.
}
