% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_BZZ_paid_per_day.R
\name{plot_BZZ_paid_per_day}
\alias{plot_BZZ_paid_per_day}
\title{Plot BZZ paid per day}
\usage{
plot_BZZ_paid_per_day(storage_per_day)
}
\arguments{
\item{storage_per_day}{Dataframe of certain structure, with columns for Date, BZZDistributedPerDay}
}
\value{
Returns a plot of BZZ paid out per day.
}
\description{
Function to plot a graph of BZZ paid to node operators per day.
}
