% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_storage_per_day.R
\name{plot_storage_per_day}
\alias{plot_storage_per_day}
\title{Plot storage bought per day}
\usage{
plot_storage_per_day(storage_per_day)
}
\arguments{
\item{storage_per_day}{Dataframe of certain structure, with columns for Date, BoughtStorageInTB}
}
\value{
Returns a plot of storage bought per day.
}
\description{
Function to plot a graph of storage paid for per day.
}
